Implement methods for {Diagonal|Mask} <-> Tensor

+++ tenpy/linalg/tensors.py
@@ -2454,9 +2454,9 @@ class Mask(AbstractTensor):
         This function determines if this action is the same."""
         raise NotImplementedError  # TODO
 
-    def to_full_tensor(self) -> Tensor:
+    def to_full_tensor(self, dtype=Dtype.float64) -> Tensor:
         return Tensor(
-            data=self.backend.full_data_from_mask(self),
+            data=self.backend.full_data_from_mask(self, dtype=dtype),
             legs=self.legs, backend=self.backend, labels=self.labels
         )
